Update main.js
I noticed that html in user's message is being rendered as html. If a user posts "<a href...>some text</a>", it is rendered as a link in userChatBox. Clicking it even crashes Acode. Apart from not being able to use my message as reference, this causes other undesirable effect such as not being able to recopy my question for reuse. 

Here, I've updated that line to display user's message as text instead of being rendered as html.